### 2024-06-03 0ecfddd
changing data format in paper.md
