### 2024-06-03 #6827 @ionides
Dear JOSS editors,

Apologies for submitting a paper.md which failed to compile.
I tested it successfully on my computer via:

docker run --rm -it  -v $PWD:/data  -u $(id -u):$(id -g)
 openjournals/inara  -o pdf,crossref paper.md

From the error message at
https://github.com/openjournals/joss-papers/actions/runs/9349651694/job/25731446811
it seemed possible that the problem with the JOSS GitHub action was related
to
the date format. On comparing with the example.md provided by JOSS, I
noticed a
difference in the date formatting. I have now changed the date format, and
pushed
a new version to the joss branch of https://github.com/kidusasfaw/spatPomp

It was unclear to me whether there is a way to run exactly the JOSS GitHub
action to further test paper.md. Reviewing the website, it seems like the
editorialbot may be able to do this. I'm experimenting with an editorialbot
command in this email to see if that helps.

Thanks,
Ed

Edward L. Ionides
Associate Chair for Undergraduate Studies and Professor,
Department of Statistics, University of Michigan
1085 South University, Ann Arbor, MI 48109-1107
email: ***@***.***
phone: 734 615 3332
office: 453 West Hall

On Mon, Jun 3, 2024 at 10:19 AM Chris Vernon ***@***.***>
wrote:

> @editorialbot <https://github.com/editorialbot> invite @RMeli
> <https://github.com/RMeli> as editor
>
> 👋 @RMeli <https://github.com/RMeli> can you take this on one as editor?
>
