### 2024-06-03 #6827 @ionides
I've pushed a version quoting all affiliation numbers in paper.md

Edward L. Ionides
Associate Chair for Undergraduate Studies and Professor,
Department of Statistics, University of Michigan
1085 South University, Ann Arbor, MI 48109-1107
email: ***@***.***
phone: 734 615 3332
office: 453 West Hall

On Mon, Jun 3, 2024 at 10:26 AM The Open Journals editorial robot <
***@***.***> wrote:

> ⚠️ An error happened when generating the pdf
> <https://github.com/openjournals/joss-papers/actions/runs/9351991353?check_suite_focus=true>.
> Problem with affiliations for Kidus Asfaw, perhaps the affiliations index
> need quoting?.
>
