### 2024-06-03 #6827 @ionides
Here are some referee suggestions:

Ben Bolker (bolker at mcmaster.ca)
https://ms.mcmaster.ca/~bolker/
An expert on spatiotemporal population biology and statistical software.
If he agrees to review, that would be great!

John Lees (jlees at ebi.ac.uk)
https://www.infectiousdisease.cam.ac.uk/staff/dr-john-lees
The corresponding author of a related open-source project,
https://wellcomeopenresearch.org/articles/5-288/v2

Len Thomas (len.thomas at st-andrews.ac.uk)
https://lenthomas.org/
Interested in partially observed Markov models in ecology, as
well as statistical software.
