### 2024-06-03 #6827 @crvernon
> @editorialbot generate pdf I've pushed a version quoting all affiliation numbers in paper.md

Thanks @ionides!
