### 2024-06-12 #6827 @gkthiruvathukal
@crvernon or @arfon: What editorialbot command will give me the list of similar papers? It seems not to have been triggered for this submission, and the "commands" command gives no insight on this. I think we need an issue to make sure this is easy to find.

Somehow, I thought "generate pdf" would trigger this. It does not.
