### 2024-06-12 #6827 @crvernon
> @crvernon or @arfon: What editorialbot command will give me the list of similar papers and identify potential reviewers? It seems not to have been triggered for this submission's pre-review, and the "commands" command gives no insight on this. I think we need an issue to make sure this is easy to find.
>
> Somehow, I thought "generate pdf" would trigger this. It does not.

:wave: @gkthiruvathukal - Let me make sure that there is a command to actually do this outside of what gets executed when the Pre-Review thread starts.  I'll do so now and come back with an answer.  Thanks!
