### 2024-06-13 #6827 @gkthiruvathukal
@crvernon Let's hope for `@editorialbot suggest-reviewers`. It would be great if we can get the usernames, when clicked, to trigger an `@editorialbot invite <username> as reviewer`.
