### 2024-06-13 #6827 @gkthiruvathukal
@crvernon: Signaling @arfon as he and I chatted about this long ago during one of our amazing calls! The fact that we have this pre-review feature is the result of an amazing pairing with Arfon. I wanted to code it but he went ahead with the implementation (ok with that!)

I would really like to eliminate the tedium of having to generate _ad hoc_ invitation requests as well. Perhaps we can have the invite semantics be to generate a comment asking whether the prospective reviewer would be willing to edit this submission. It would be really amazing if they could just click a yes/no checkbox and be self-assigned.

I wonder whether I should open an issue, if only to articulate the thinking, which is based on many years of EIC service in IEEE publications. Most publications have workflows with few (if any) manual steps. AEICs (like you) and Editors (like me) should be able to have personalized templates for various requests, e.g. inviting reviewers, nudging reviewers, and checklists for acceptance.

Meanwhile, I'm going to go with @ionides's suggestions!
