### 2024-07-13 #6827 @gkthiruvathukal
@bbolker, @lenthomas, and @johnlees: Are you available to contribute a review for this JOSS submission? I need at least two reviewers to proceed.
