### 2024-07-18 #7008 @editorialbot
[REVIEW]: spatPomp: An R package for spatiotemporal partially observed Markov process models
**Submitting author:** @ionides (<a href="http://orcid.org/0000-0002-4190-0174">Edward Ionides</a>)
**Repository:** https://github.com/kidusasfaw/spatPomp
**Branch with paper.md** (empty if default branch): joss
**Version:** 1.0.0
