### 2024-07-30 kidusasfaw/spatPomp#26 @johnlees
Possible improvements to example usage
Current usage examples are detailed statistical PDFs, or source code. An improvement to users would be to consider adding one of these as a formal R package vignettes, at least for one of the listed resources. I think a simple example running the iterated block particle filter would be nice.

I was able to follow through the existing PDF (which has a lot of troubleshooting and plotting which is great), but some commands didn't work 'out of the box' e.g.:

```r
Warning message:
In seq.default(from = 0.2, to = 0.8, length = 4) :
  partial argument match of 'length' to 'length.out'
```
```r
Error: object 'U' not found
```

And the doParallel setup was not explicit
