### 2024-07-30 kidusasfaw/spatPomp#27 @johnlees
Add CONTRIBUTING
For JOSS, add a community contributions guide, either as `.github/CONTRIBUTING` or in the README. This should contain info on:
1) How to contribute to the software
2) How to report issues or problems with the software
3) How to seek support
