### 2024-07-30 #7008 @johnlees
This is a useful piece of software that is a substantial contribution to the field. I have the following comments on the checklist:

> Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

See https://github.com/kidusasfaw/spatPomp/issues/26

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

See https://github.com/kidusasfaw/spatPomp/issues/27

> Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

The summary is currently very heavy on technical, statistical language. Leading with the example paragraph and giving more details on a use case would perhaps help. 'metapopulations' and 'unit structure' should be defined to non-experts. Changing the order of sentences in the first paragraph would also be sensible, it feels like the order is almost backwards at the moment.

> State of the field: Do the authors describe how this software compares to other commonly-used packages?

This is partially done but only with pomp. It would be useful to a general audience why other packages in this domain (e.g. libBi and odin.dust) and more general statistical packages (e.g. stan) don't fulfill the needs that spatPomp does
