### 2024-08-03 646e270
Create CONTRIBUTING.md
Instructions for open-source contributors to raise issues and update the code base.
