+++ CONTRIBUTING.md
@@ -1,0 +1,51 @@
+Thank you for your interest in improving our package. Here are some instructions to ensure that we are aware of your thoughts and contributions. These templates are inspired by those in the [pomp](https://github.com/kingaa/pomp/tree/master/.github/ISSUE_TEMPLATE) package.
+### Bug reports
+To report a bug please open an issue ticket and follow the template below:
+
+**Describe the bug**\
+Please provide a clear and concise description of the bug.
+
+**To Reproduce**
+1. **R** code that produces the problem.
+2. Transcript of the R session that gave the errors.
+3. Any graphics that show the problematic results.
+
+**Expected behavior**\
+A clear and concise description of what you expected to happen.
+
+**Session information**\
+Output of the command
+```
+source("https://kingaa.github.io/scripts/diagnostics.R")
+```
+
+### Feature requests
+To request a feature, please open an issue ticket and follow the template below:
+
+**Is your feature request related to a problem? Please describe.**\
+A clear and concise description of what the problem is. Ex. I'm always frustrated when [...]
+
+**Describe the solution you'd like**\
+A clear and concise description of what you want to happen.
+
+**Describe alternatives you've considered**\
+A clear and concise description of any alternative solutions or features you've considered.
+
+**Additional context**\
+Add any other context or screenshots about the feature request here.
+
+### Creating a pull request
+Please use the following template for your pull request to help us understand the need for the change and whether it is ready to be integrated in the code base.
+
+**Issue**\
+Is your problem addressing an existing issue ticket? If so please link to it. Otherwise, please open one first and link to it here.
+
+**File changes**\
+Please provide a detailed description in words of the changes to each file.
+
+**Unit tests**\
+Please ensure that you have run all existing unit tests. In addition, it is likely that you will need to add your own unit tests to test your code changes.
+Please check the boxes that apply:
+
+- [ ] I have run the existing unit tests
+- [ ] I have added unit tests for my changes
