### 2024-08-03 kidusasfaw/spatPomp#27 @kidusasfaw
Hi @johnlees, thank you for pointing out this improvement to our package. I have now added a CONTRIBUTING markdown file to the root directory.
