### 2024-08-06 db3e526
update tutorials and references
