### 2024-08-06 b2bdf3d
try adding tinytex to github action to fix check
