### 2024-08-06 15c3711
add latex packages for tinytex github action
