+++ .github/workflows/r-cmd-check.yml
@@ -52,9 +52,18 @@ jobs:
       _R_CHECK_CODETOOLS_PROFILE_: "suppressLocalUnused=FALSE,suppressPartialMatchArgs=FALSE,suppressParamUnused=TRUE,suppressUndefined=FALSE"
 
     steps:
-      - uses: r-lib/actions/setup-tinytex@master
       
       - uses: actions/checkout@v4
+
+      - uses: r-lib/actions/setup-tinytex@v2
+      - run: |
+          tlmgr --version
+          tlmgr update --self
+          tlmgr install titlesec
+          tlmgr install pdfpages
+          tlmgr install pdflscape
+          tlmgr install ae
+          tlmgr list --only-installed
 
       - uses: r-lib/actions/setup-r@v2
         with:
