### 2024-08-06 2e0cc65
fix doi links and start addressing peer review
