+++ paper.bib
@@ -3,7 +3,7 @@
   title={A tutorial on spatiotemporal partially observed {M}arkov process models via the {R} package spat{P}omp},
   author={Asfaw, Kidus and Park, Joonha and King, Aaron A and Ionides, Edward L},
   journal={arXiv:2101.01157v4},
-  url={https://arxiv.org/abs/2101.01157},
+  doi={10.48550/arXiv.2101.01157},
   year={2024}
 }
 
@@ -11,14 +11,16 @@
   title={An introduction to sequential {M}onte {C}arlo},
   author={Chopin, Nicolas and Papaspiliopoulos, Omiros},
   year={2020},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/978-3-030-47845-2}
 }
 
 @book{evensen22,
   title={Data assimilation fundamentals: A unified formulation of the state and parameter estimation problem},
   author={Evensen, Geir and Vossepoel, Femke C and Van Leeuwen, Peter Jan},
   year={2022},
-  publisher={Springer Nature}
+  publisher={Springer Nature},
+  doi={10.1007/978-3-030-96709-3}
 }
 
 @article{ionides23,
@@ -31,22 +33,14 @@
   doi = {10.1080/01621459.2021.1974867}
 }
 
-@misc{ionides23code,
-  author = {Edward L. Ionides and Kidus Asfaw and Joonha Park and Aaron A. King},
-  title = {Source code for ``Bagged filters for partially observed interacting systems''},
-  year = {2023},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/ionides/bagged_filters}
-}
-
 @article{kalman60, 
-author = "Kalman, R. E.",
-year = 1960, 
-title = "A new approach to linear filtering and prediction problems",
-journal = "Journal of Basic Engineering",
-volume = 82, 
-pages = "35--45"
+  author = {Kalman, R. E.},
+  year = 1960, 
+  title = {A new approach to linear filtering and prediction problems},
+  journal = {Journal of Basic Engineering},
+  volume = 82, 
+  pages = {35--45},
+  doi={10.1115/1.366255}
 }
 
 @Article{king16,
@@ -60,19 +54,14 @@ pages = "35--45"
 }
 
 @article{li24,
-  title={Inference on spatiotemporal dynamics for networks of biological populations},
+  title={Inference on spatiotemporal dynamics for coupled biological populations},
   author={Li, Jifan and Ionides, Edward L and King, Aaron A and Pascual, Mercedes and Ning, Ning},
-  journal={arXiv:2311.06702},
+  journal={Journal of the Royal Society Interface},
+  doi={10.1098/rsif.2024.0217},
+  volume=21,
+  number=216,
+  pages= 20240217,
   year={2024}
-}
-
-@misc{li24code,
-  title={Code for ``Inference on spatiotemporal dynamics for networks of biological populations''},
-  author={Li, Jifan and Ionides, Edward L and King, Aaron A and Pascual, Mercedes and Ning, Ning},
-  year = {2024},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/jifanli/metapop_article}
 }
 
 @article{ning23,
@@ -80,23 +69,17 @@ pages = "35--45"
   author={Ning, Ning and Ionides, Edward L},
   journal={Journal of Machine Learning Research},
   url={https://www.jmlr.org/papers/volume24/21-1253/21-1253.pdf},
+  doi={10.48550/arXiv.2110.10745},
   volume={24},
   pages={1--76},
   year={2023}
-}
-
-@misc{ning23tutorial,
[…]
