### 2024-08-07 61e0bdf
add R scripts for tutorials
