+++ vignettes/spatPomp.R
@@ -1,0 +1,580 @@
+## ----packages,include=F,echo=F,cache=F----------------------------------------
+library("spatPomp")
+library("gridExtra")
+library("grid")
+library("ggplot2")
+library("xtable")
+library("tidyverse")
+library("cowplot")
+library("knitr")
+library("kableExtra")
+
+
+
+
+
+
+## ----knitr-opts-for-purl,include=F,cache=F,purl=T-----------------------------
+opts_chunk$set(
+    warning=FALSE,
+    message=FALSE
+)
+
+
+## ----setup,cache=F,include=F--------------------------------------------------
+myround<- function (x, digits = 1) {
+  # adapted from the broman package
+  # solves the bug that round() kills significant trailing zeros
+  if (length(digits) > 1) {
+    digits <- digits[1]
+    warning("Using only digits[1]")
+  }
+  if (digits < 1) {
+    as.character(round(x,digits))
+  } else {
+    tmp <- sprintf(paste("%.", digits, "f", sep = ""), x)
+    zero <- paste0("0.", paste(rep("0", digits), collapse = ""))
+    tmp[tmp == paste0("-", zero)] <- zero
+    tmp
+  }
+}
+mysignif <- function (x, digits = 1) {
+  myround(x, digits - ceiling(log10(abs(x))))
+}
+
+
+## ----set-seed,eval=T,echo=F---------------------------------------------------
+bm_sim_U <- 10
+bm_sim_N <- 20
+set.seed(76)
+
+## ----example_bm4,eval=T,echo=F------------------------------------------------
+bm10 <- bm(U = 10, N = 20)
+
+
+## ----bm10-simplot,eval=T,echo=F,fig.height=2.5,fig.width=6.5,out.width="100%",fig.cap=paste0(r"{Result of executing \code{plot(bm10)}, where \code{bm10} is the \class{spatPomp} object representing a simulation from a }",bm_sim_U,r"{-dimensional correlated Brownian motions model with }",bm_sim_N,r"{ observations that are one unit time apart (see text).}")----
+plot(bm10,nrow=2)+
+  theme(plot.margin=unit(c(0,0.1,0,0),"cm"))
+
+
+
+
+## ----abf_parallel,eval=T,echo=T-----------------------------------------------
+library("doParallel")
+registerDoParallel(detectCores()) 
+
+## ----cores,eval=T,echo=F------------------------------------------------------
+cores <- detectCores()
+
+
+## ----example_nbhd-------------------------------------------------------------
+example_nbhd <- function(object, unit, time){
+  nbhd_list = list()
+  if(time>1) nbhd_list <- c(nbhd_list, list(c(unit, time-1)))
+  if(unit>1) nbhd_list <- c(nbhd_list, list(c(unit-1, time)))
+  return(nbhd_list)
+}
+
+
+## ----bm10-params,echo=T-------------------------------------------------------
+coef(bm10)
+
+
+
+
+## ----bm_set_i,cache=F,echo=F,eval=T-------------------------------------------
+i <- 2
+
+
+## ----bm-settings,cache=FALSE,echo=F-------------------------------------------
+bm_nbhd <- function(object, time, unit) {
[…]
