+++ vignettes/ibpf.R
@@ -1,0 +1,683 @@
+## ----cores,echo=F,cache=F,message = FALSE, warnings = FALSE,results='hide'----
+library(doParallel)
+cores <-  as.numeric(Sys.getenv('SLURM_NTASKS_PER_NODE', unset=NA))
+if(is.na(cores)) cores <- detectCores()
+registerDoParallel(cores)
+ggplot2::theme_set(ggplot2::theme_bw())
+
+
+## ----packages,include=F,echo=F,cache=F----------------------------------------
+library("spatPomp")
+library("ggplot2")
+library("tidyverse")
+library("knitr")
+stopifnot(packageVersion("pomp")>="5.0")
+
+
+
+
+
+
+## ----setup,cache=F,include=F--------------------------------------------------
+myround<- function (x, digits = 1) {
+  # adapted from the broman package
+  # solves the bug that round() kills significant trailing zeros
+  if (length(digits) > 1) {
+    digits <- digits[1]
+    warning("Using only digits[1]")
+  }
+  if (digits < 1) {
+    as.character(round(x,digits))
+  } else {
+    tmp <- sprintf(paste("%.", digits, "f", sep = ""), x)
+    zero <- paste0("0.", paste(rep("0", digits), collapse = ""))
+    tmp[tmp == paste0("-", zero)] <- zero
+    tmp
+  }
+}
+mysignif <- function (x, digits = 1) {
+  myround(x, digits - ceiling(log10(abs(x))))
+}
+
+
+## ----bm-setup,echo=F,eval=T---------------------------------------------------
+set.seed(20)
+
+
+## ----bm-model,echo=T,eval=T---------------------------------------------------
+library(spatPomp)
+i <- 2
+U <- 4
+b <- bm2(U=U,N=switch(i,10,200),unit_specific_names="rho")
+plot(b)
+
+
+## ----bm-followup,echo=F,eval=T------------------------------------------------
+if(i==1) cores <- 2   # no need to use all available cores when i=1
+
+
+
+
+## ----bm_logLik_eval,cache=F,echo=F,eval=T-------------------------------------
+bm_dir <- paste0("bm_",i,"/")
+if(!dir.exists(bm_dir)) dir.create(bm_dir)
+stew(file=paste0(bm_dir,"logLik.rda"),seed=10,{
+
+  cat(capture.output(sessionInfo()),
+    file=paste0(bm_dir,"sessionInfo.txt"),sep="\n")
+
+kf_logLik <- bm2_kalman_logLik(b)
+pf_logLik <- replicate(10,
+  logLik(pfilter(b,switch(i,10,1000)))
+)
+bpf_logLik2 <- replicate(10,
+  logLik(bpfilter(b,switch(i,10,1000),block_size=2))
+)
+  bpf_logLik1 <- replicate(10,
+    logLik(bpfilter(b,switch(i,10,1000),block_size=1))
+  )
+  bpf_logLik4 <- replicate(10,
+    logLik(bpfilter(b,switch(i,10,1000),block_size=4))
+  )
+})
+
+
+## ----kf_mle,echo=F,eval=T-----------------------------------------------------
+stew(file=paste0(bm_dir,"kf_mle.rda"),seed=999,{
[…]
