### 2024-08-08 kidusasfaw/spatPomp#26 @johnlees
Excellent, thank you for these comprehensive changes! I will update the JOSS review
