### 2024-08-08 kidusasfaw/spatPomp#27 @johnlees
Great, I will update the JOSS review
