### 2024-08-08 #7008 @johnlees
Issues https://github.com/kidusasfaw/spatPomp/issues/26 and https://github.com/kidusasfaw/spatPomp/issues/27 addressed and I have marked those boxes as complete accordingly
