### 2024-08-08 0ae3a14
including feedback on paper.pdf from john lees
