+++ paper.bib
@@ -23,6 +23,16 @@
   doi={10.1007/978-3-030-96709-3}
 }
 
+@article{fitzjohn20,
+  title={Reproducible parallel inference and simulation of stochastic state space models using odin, dust, and mcstate},
+  author={FitzJohn, Richard G and Knock, Edward S and Whittles, Lilith K and Perez-Guzman, Pablo N and Bhatia, Sangeeta and Guntoro, Fernando and Watson, Oliver J and Whittaker, Charles and Ferguson, Neil M and Cori, Anne and Baguelin, Marc and Lees, John A},
+  journal={Wellcome Open Research},
+  volume={5},
+  year={2020},
+  doi={10.12688/wellcomeopenres.16466.2},
+  publisher={The Wellcome Trust}
+}
+
 @article{ionides23,
   author = {Edward L. Ionides and Kidus Asfaw and Joonha Park and Aaron A. King},
   title = {Bagged filters for partially observed interacting systems},
@@ -53,6 +63,18 @@
   doi     = {10.18637/jss.v069.i12}
 }
 
+@article{li18,
+  title={Fitting mechanistic epidemic models to data: A comparison of simple {M}arkov chain {M}onte {C}arlo approaches},
+  author={Li, Michael and Dushoff, Jonathan and Bolker, Benjamin M},
+  journal={Statistical Methods in Medical Research},
+  volume={27},
+  number={7},
+  pages={1956--1967},
+  year={2018},
+  doi={10.1177/0962280217747054},
+  publisher={SAGE Publications Sage UK: London, England}
+}
+
 @article{li24,
   title={Inference on spatiotemporal dynamics for coupled biological populations},
   author={Li, Jifan and Ionides, Edward L and King, Aaron A and Pascual, Mercedes and Ning, Ning},
@@ -62,6 +84,28 @@
   number=216,
   pages= 20240217,
   year={2024}
+}
+
+@article{michaud21,
+  title={Sequential {M}onte {C}arlo methods in the nimble and {nimbleSMC} {R} Packages},
+  doi={10.18637/jss.v100.i03},
+  author={Michaud, Nicholas and de Valpine, Perry and Turek, Daniel and Paciorek, Christopher J and Nguyen, Dao},
+  journal={Journal of Statistical Software},
+  volume={100},
+  pages={1--39},
+  year={2021}
+}
+
+@article{murray15,
+ title={Bayesian state-space modelling on high-performance hardware using {LibBi}},
+ volume={67},
+ xurl={https://www.jstatsoft.org/index.php/jss/article/view/v067i10},
+ doi={10.18637/jss.v067.i10},
+ number={10},
+ journal={Journal of Statistical Software},
+ author={Murray, Lawrence M.},
+ year={2015},
+ pages={1–36}
 }
 
 @article{ning23,
