### 2024-08-08 #7008 @ionides
@johnlees, thanks again for the feedback on the spatPomp code and manuscript. I'm unsure of the protocols for the open review process at JOSS, but it seems to make sense to respond now to your two points about the manuscript. The revised version is available as paper.pdf on the joss branch of the spatPomp repo (https://github.com/kidusasfaw/spatPomp/blob/joss/paper.pdf).

1. _The summary is currently very heavy on technical, statistical language. Leading with the example paragraph and giving more details on a use case would perhaps help. 'metapopulations' and 'unit structure' should be defined to non-experts. Changing the order of sentences in the first paragraph would also be sensible, it feels like the order is almost backwards at the moment._

We have rewritten the summary using this feedback. It now starts with introducing a concrete application and moves toward the general situation, whereas the original submission took the reverse order.

2. _It would be useful to a general audience why other packages in this domain (e.g. libBi and odin.dust) and more general statistical packages (e.g. stan) don't fulfill the needs that spatPomp does. The sentence 'We are not aware of alternative packages providing statistically efficient, plug-and-play inference for the general class of SpatPOMP models.' would be improved by citing some existing packages and explaining what they do/don't do instead._

Thank you for pointing out this oversight. Additional discussion of alternative packages is in the extended package tutorial but we agree that this topic should be included also in the relatively brief JOSS overview. We have added some more sentences to the paragraph on the relationship to previous packages, which now reads as follows:

The `spatPomp` package builds on `pomp` (King et al. 2016) which is a successful software package for low-dimensional POMP models. Other packages with similar capabilities to `pomp` include `nimble` (Michaud et al. 2021), `LiBBi` (Murray et al. 2015) and `mcstate` with `odin` and `dust` (FitzJohn et al. 2020). All these packages enable plug-and-play inference based on sequential Monte~Carlo. Markov chain Monte Carlo packages, such as `stan`, have been found to be effective for inference on some POMP models (Li et al. 2018) but they lack the plug-and-play property. Perhaps for that reason, sequential Monte Carlo methods have found broader applicability for this model class. We are not aware of alternative packages to `spatPomp` that provide statistically efficient, plug-and-play inference for the general class of SpatPOMP models.
