### 2024-08-09 #7008 @johnlees
@ionides Thanks for these changes, I think both sections read very well. Thanks also for the small tweaks to the package you already addressed in the issues.

Congratulations and thank you for another useful addition to software packages in this important area. I have marked all the other boxes in my review checklist as complete.
