### 2024-10-21 #7008 @bbolker
Thank you for your extreme patience. The package looks fine to me, the vignette is extremely clear and detailed. A few general comments:

* it's moderately hard to do non-trivial checks on packages where most of the computation is non-trivial. The package does have some checks. It might be useful in the long run to look at porting tests to `tinytest`, which can more easily do granular testing.
* glad there's a prior/regularization possibility (and a link to the `pomp` documentation
* What if initial values are estimated from dynamical parameters (as in "model starts from equilibrium density of an SIR model") ? (theta_IVP vs theta_RP) ?
* it would be nice if there were syntax tricks to make snippets indentable/highlightable, but this is really a more general `pomp` question

## details

These are mostly comments on the vignette.

* Is there a reason that `simulate(..., format = "data.frame")` gives unit name `unitname`; `as.data.frame(simulate(...))` gives unit name `unit`?
* section 2.8, avoid smart quotes in `format = "data.frame"`
* are extended filters (GIRF, APF etc.) useful/available for non-spatial problems too?
* "platform for developing new statistical methods" - by authors, or anyone?  If the intention is to make this a general platform (analogous to NIMBLE), it might be useful to have developer guidelines.
* I wondered about the possibility of including synthetic likelihood (see Fasiolo and Wood [synlik package](https://mfasiolo.github.io/synlik_release/)) as an alternative method (most similar to ABC, although dependent on being able to model the multivariate distribution of summary statistics as a multivariate Gaussian or some extension thereof (e.g. multivariate skew-t)

The setup of the package and GitHub repository make it a little bit harder than I'd like to access the underlying materials. For example, the vignettes are understandably set up as shells to import a pre-compiled PDF.  That makes sense because the vignette computations would certainly be too slow for testing on CRAN, but where are the vignette sources? They could be included in `inst/` somewhere (and a Makefile could be included for building the vignettes from scratch, or at least from stored `rds`/`rda` files containing the results of computationally expensive fits ...)  I list a bunch of small typos in the vignette below - if the vignette source had been on the repo I could have fixed them myself and submitted a pull request ...

## code details

* it might be useful to have a choice of tracing frequency/verbosity  for the `verbose` option
* might make sense to rename the 'verbose' option to something package-specific (e.g. `options("spatPomp_verbose")`), to avoid potential classes with other packages using the same kind of mechanism

## tutorial vignette details

* In general I prefer to see log-likelihoods stated on some scale relative to the best (reference) model (i.e. `rel_llik = max(likvec) - likvec`, or the opposite for negative log-likelihoods; in the case of Figure 4, that would give values relative to KF (e.g. p. 24)
* possibly consider new-style code formatting in the vignette (no prompts or continuation markers) to make cutting and pasting easier
* Figure 5: show loglik on a log scale from final/max value? (i.e., the large increase in log-likelihood from its initial value makes it harder to distinguish changes during the plateau - showing the log of relative log-likelihood may seem a little redundant but makes small changes easier to see ...)
* p. 25 "contains a parameter estimate [and] its log-likelihood"
* p. 27 "are counting process[es]"
* p. 29 "We [construct] the model"
* p. 30 could single quotes in code be type set as straight quotes ' rather than single back-quotes ´  ?
* p. 32 not clear what the "parameter vector of scientific interest" is ...
* R0 = 30 seems high for measles (I do know that some studies have estimated it as that high or higher ...)
* p. 35 italicize Leptospira interrogans ?

## ibpf vignette

* p.7 "provide[s] an adequate"
* p. 7 delete comma after "but"
* "Jensen's [inequality]"
* "This is [in] contrast"
* p. 9 "for one [parameter]"
* p. 12 "agree on the most problematic observation[s]" ?
* p. 14 "For our measles spatPomp [example]"
* Fig 8: are the trace plots constant for the last parameters since these were held fixed in the example? Is it true that we shouldn't interpret the trace plots in quite the same way as for MCMC (since we are not trying to take a sample, i.e. we don't care about mixing or burn-in)?
* Wheeler et al;  "Ionides, E. L." (not j."

## other points

* `plot` method could/should add a y-axis label to the plot that indicates the log10(x+1) transformation? (Why not use `scale_y_log10()` [which automatically puts zero values (log(x) → -∞) at the edge of the plot] or the `log1p` transformation?
* titles of some help pages are noninformative (`runit_measure`, `munit_measure`)
* 'plot' page headings could be improved
* use FALSE, TRUE rather than T, F;  clarify that log=TRUE uses log(1+x)
