### 2024-10-22 #7008 @bbolker
I have completed the checklist and given some suggestions (none of which are necessary to address before considering package acceptable). Is there anything else I need to do to move this (very belated because of me) process along to the next stage?
