### 2024-10-23 #7008 @gkthiruvathukal
@bbolker Thanks for your detailed feedback. This is wonderful!

@ionides Can you work to address as many of these as possible and follow up when done? As @bbolker mentioned, these are not blockers for acceptance. Still, we'd like to make sure your submission has maximal impact, which in part will be achieved by making sure your submission is as free of issues as possible.

I'd like to move this submission forward as soon as possible, so please follow up and I can direct you to the next steps.
