### 2024-10-23 #7008 @ionides
Thank you, @bbolker and @gkthiruvathukal. It will probably take a couple of weeks to incorporate Ben's thoughtful suggestions, and at that point I'll get back to you.
