+++ R/plot.R
@@ -1,69 +1,46 @@
-##' Plotting \code{spatPomp} data
+##' Plot methods for \code{spatPomp} objects
 ##'
-##' Visualize \code{spatPomp} data
+##' Visualize data in a \code{spatPomp} object or a derived class.
+##' This gives a quick view; the data can be extracted from 
+##' the object to make a customized plot.
+##'
 ##' @name plot
 ##' @rdname plot
 ##' @include spatPomp_class.R
 ##' @param x a \code{spatPomp} object
-##' @param log should the data be log-transformed before plotting?
+##' @param log should the data be transformed to \code{log10(x+1)} before plotting?
 ##' This helps in contexts where there are spikes that could take away
-##' attention from the dynamics illustrated by the rest of the data.
+##' attention from the dynamics illustrated by the rest of the data. 
 ##' @param ncol the number of columns in the grid plot
 ##' @param type for visualizing an object of class \code{spatPomp}, the user
-##' can obtain a grid of line plots by default (\code{'l'}) or a heat map by
-##' supplying argument \code{'h'}.
+##' can obtain a grid of line plots by default (\code{l}) or a heat map by
+##' supplying argument \code{h}.
 ##' @param plot_unit_names allows suppression of unit names when making
 ##' a heat map for a large number of units
+##' @param params allows selection of a subset of parameters when making
+##' a diagnostic plot for a model with many parameters
 ##' @param ...  for visualizing an object of class \code{spatPomp}, the user
 ##' can add arguments like \code{nrow} to specify the number of rows in the
 ##' grid.
+##' @return a \code{ggplot} plot of class \sQuote{gg} and \sQuote{ggplot}.
+##' @importFrom ggplot2 ggplot geom_line aes facet_wrap
+##' @importFrom graphics par
 NULL
 
 if(getRversion() >= "2.15.1")  utils::globalVariables(c("."))
-##' Plotting output of \code{igirf()}
-##'
-##' Diagnostic plot for \code{igirf()}
-##' @param params the names of the parameters for which the user would like to see a trace plot
-##' @name plot-igirfd_spatPomp
-##' @rdname plot
-##' @importFrom ggplot2 ggplot geom_line aes facet_wrap
-##' @importFrom graphics par
-##' @aliases plot,igirfd_spatPomp-method
-##' @return a \code{ggplot} facet plot of class \sQuote{gg} and \sQuote{ggplot} visualizing
-##' the convergence record of running \code{igirf()} with respect to the likelihood and the parameters of the model.
-setMethod(
-  "plot",
-  signature=signature(x="igirfd_spatPomp"),
-  definition=function (x, params = names(coef(x)), ncol = 3) {
-    plot.df <- data.frame(x@traces[,c("loglik", params)])
-    cn <- colnames(plot.df)
-    plot.df <- cbind(c(seq_len(dim(plot.df)[1])), plot.df)
-    names(plot.df) <- c("iteration", cn)
-    to.gather <- colnames(plot.df)[2:length(colnames(plot.df))]
-    to.plot <- plot.df |> tidyr::gather(key = "param", val = "value", tidyr::all_of(to.gather)) |> tail(-1)
-    ggplot2::ggplot(data = to.plot) +
-      ggplot2::geom_line(mapping = ggplot2::aes(x = .data$iteration, y = .data$value)) +
-      ggplot2::facet_wrap(~param, ncol = ncol, scales = "free")
-  }
[…]
