### 2024-10-29 #7008 @ionides
Hi @bbolker, Thanks for your detailed feedback. Here is a point-by-point response, with your text in bold. Changes to the JOSS manuscript, paper.pdf, and the spatPomp package are pushed to the github repository. Package changes will not appear on CRAN until the next upload.

__1. Thank you for your extreme patience. The package looks fine to me, the vignette is extremely clear and detailed.__

* Your thoughtful comments are worth waiting for!

### General comments

__2. It's moderately hard to do non-trivial checks on packages where most of the computation is non-trivial. The package does have some checks. It might be useful in the long run to look at porting tests to tinytest, which can more easily do granular testing.__

* Thanks for pointing out tinytest which looks like an attractive alternative to testthat. The test approach for spatPomp follows that of pomp, and is suited to the task at hand though it uses only the base R unit-test setup. The unit tests focus on fine-grain tests by having 100% line coverage. At a coarse granularity, the tests check whether results of high-level functions have changed, but do not check whether those results are correct. Change checks are made by comparing relatively quick calculations with a fixed seed. The correctness checks are carried out by exhaustive Monte Carlo calculation test cases with results that can be validated by exact calculation. These computationally intensive tests are carried out by the tutorials, which are updated less frequently. Usually, a change check is enough, but occasionally it should be re-checked against an absolute standard. At this point, there is no urgency to refactor the unit tests.

__3. Glad there's a prior/regularization possibility (and a link to the pomp documentation)__

* Good to hear that it is appreciated.

__4. What if initial values are estimated from dynamical parameters (as in "model starts from equilibrium density of an SIR model") ? (theta_IVP vs theta_RP) ?__

* This question is valid for all POMP models, but there's no reason why it shouldn't be clarified in our particular SpatPOMP context.
We have added the following to Section 2.4 (Initial conditions) of the tutorial article.

_For a stationary model, the initial state may be set at the stationary distribution by specifying $t_0$ sufficiently remote from $t_1$ to allow the system to equilibrate.
Stocks et al. (2020) provides an example of this using pomp.
When following this approach, the model is insensitive to the choice of $f_{X_0}(x_0| \theta)$._

* In addition, we have noted:

_The RPs are also provided to `rinit` so they can participate in the initialization._

__5. It would be nice if there were syntax tricks to make snippets indentable/highlightable, but this is really a more general pomp question__

* We understand this as a suggestion to develop setup options for coding environments such as Rstudio or emacs or vscode which could enable them to format the C code in Csnippets, adapting automatically to the need for different syntax highlighting of R and C sections. You are correct that this should first be incorporated into pomp, so it is not an immediate spatPomp development task.

### Details

__These are mostly comments on the vignette.__

__6. Is there a reason that `simulate(..., format = "data.frame")` gives unit name `unitname`; `as.data.frame(simulate(...))` gives unit name `unit`?__

* One might reasonably expect `as.data.frame(simulate(spatPomp_object))` to give the same output as `simulate(spatPomp_object,format="data.frame")`. However, these alternative approaches arise from different motivations that lead to different results. `simulate(format="data.frame")` is used as a low-level function within functions such as `abc` or `probe.match`.  It should quickly be able to simulate many realizations with argument `nsim>1`.  By contrast, `as.data.frame(spatPomp_object)` is used to inspect all the data in a single pomp object and produce custom plots.

* The potentially surprising lack of commutativity when `nsim=1` is inherited from pomp, and is quite a deep property of the package so we are not changing it at this time.

__7. Section 2.8, avoid smart quotes in `format = "data.frame"`__

* Done.

__8. Are extended filters (GIRF, APF etc.) useful/available for non-spatial problems too?__

* Yes, in principle they are applicable to general complex high-dimensional models which are not SpatPOMPs. That is outside the scope of spatPomp.

__9. "platform for developing new statistical methods" - by authors, or anyone? If the intention is to make this a general platform (analogous to NIMBLE), it might be useful to have developer guidelines.__

* All are welcome to contribute new functionality to the package, via the standard GitHub interface. Some researchers have already contributed, and we look forward to more in future. The tutorial article contains a discussion of the spatPomp API which is a good foundation for prospective developers. To clarify this, we have added a comment in the JOSS article discussion of this tutorial:

_Section 2 explains the elementary methods used to access properties of the `spatPomp` model class.
These elementary methods are the building blocks available to developers for implementing complex algorithms acting on spatPomp models._

* In addition, the spatPomp package has a CONTRIBUTING.md file, though this is aimed primarily at guiding bug-fix and feature-request contributions rather than substantial development contributions.

__10. I wondered about the possibility of including synthetic likelihood (see Fasiolo and Wood synlik package) as an alternative method (most similar to ABC, although dependent on being able to model the multivariate distribution of summary statistics as a multivariate Gaussian or some extension thereof (e.g. multivariate skew-t)__

* This is currently possible using the pomp function `probe.match`, taking advantage of the property that the spatPomp class inherits from pomp. Probe matching does not scale well to a large number of parameters or a large number of probes (the name given in pomp to the statistics modeled as multivariate normal in Wood's synthetic likelihood approach). That is why spatPomp has not focused on probe matching and synthetic likelihood.

* Synthetic likelihood, used in a Bayesian approach, can also be viewed as a generalization of approximate Bayesian computing (ABC). We discuss ABC briefly, in Section 4.5 of the tutorial. We have added a mention to synthetic likelihood at this point.

__11. The setup of the package and GitHub repository make it a little bit harder than I'd like to access the underlying materials. For example, the vignettes are understandably set up as shells to import a pre-compiled PDF. That makes sense because the vignette computations would certainly be too slow for testing on CRAN, but where are the vignette sources? They could be included in inst/ somewhere (and a Makefile could be included for building the vignettes from scratch, or at least from stored rds/rda files containing the results of computationally expensive fits ...) I list a bunch of small typos in the vignette below - if the vignette source had been on the repo I could have fixed them myself and submitted a pull request$\dots$__

* The tutorial vignette has its own git repository, at [https://github.com/ionides/spatPomp-article](https://github.com/ionides/spatPomp-article). This is linked from the arXiv version of the tutorial and the vignette version.

* The iterated block particle filtering vignette is developed at [https://github.com/ionides/ibpf-tutorial](https://github.com/ionides/ibpf-tutorial). This is linked from the arXiv version of the tutorial and the vignette version.

* We have added a README to spatPomp/vignettes so that anyone disappointed at not finding the source code there will be immediately directed to the right place.

* There are advantages and disadvantages to having separate repos for long and computationally involved vignettes.
The development of the package code is cleanest when separated from related add-ons, and that is the main reason for our decision.

## Code details

__12. It might be useful to have a choice of tracing frequency/verbosity for the verbose option__

* spatPomp so far has focused on likelihood optimization rather than posterior calculation using Markov chain Monte Carlo (MCMC). Effective approaches to MCMC for general SpatPOMP models have not yet become established. For small POMP models, particle MCMC is generally applicable and has become established as an effective method. However, particle MCMC scales poorly and is not a serious contender for SpatPOMP models. Therefore, our 'traces' correspond to order 100 optimization steps rather than much larger numbers (say, $10^5$) of MCMC steps. We are postponing the treatment of long traces until the issue materializes.

__13. Might make sense to rename the 'verbose' option to something package-specific (e.g. options("spatPomp_verbose")), to avoid potential classes with other packages using the same kind of mechanism__

* Done.

### Tutorial vignette details

__14. In general I prefer to see log-likelihoods stated on some scale relative to the best (reference) model (i.e. `rel_llik = max(likvec) - likvec`, or the opposite for negative log-likelihoods; in the case of Figure 4, that would give values relative to KF (e.g. p. 24)__

* We agree that log-likelihoods are almost always interpreted via differences. However, we have found it useful to consistently report the absolute log-likelihood rather than a relative log-likelihood. When comparing many models and methods for the same dataset, it can be unclear which is the appropriate reference value. The absolute log-likelihood for a given dataset has a nice property that it can be compared quickly against some new model. As long as the reference likelihood is provided explicitly, one can retrieve the absolute log-likelihood, but it is an awkward way to provide the information if the reader wants to make their own comparisons. We understand that there are valid arguments in favor of both approaches, and we have chosen to write the tutorial consistent with our favored approach.

__15. possibly consider new-style code formatting in the vignette (no prompts or continuation markers) to make cutting and pasting easier__

* This is a good idea and we have implemented it. Thank you!

__16. Figure 5: show loglik on a log scale from final/max value? (i.e., the large increase in log-likelihood from its initial value makes it harder to distinguish changes during the plateau - showing the log of relative log-likelihood may seem a little redundant but makes small changes easier to see ...)__

* We agree that relevant details on this likelihood plot are inapparent on this scale. Here, we are demonstrating the default diagnostic plot which has a simple generic behavior. For the revision, we have also explained to the reader how to access the data for making a plot customized for their particular analysis.

__17. p. 25 "contains a parameter estimate [and] its log-likelihood"__

Done.

__18. p. 27 "are counting process[es]"__

Done.

__19. p. 29 "We [construct] the model"__

Done.

__20. p. 30 could single quotes in code be type set as straight quotes ' rather than single back-quotes ´?__

Done.

__21. p. 32 not clear what the "parameter vector of scientific interest" is ...__

A fair point. We meant to make a general assertion that the user will probably put in a parameter vector that seems meaningful to them, but we've deleted "of scientific interest" since it is redundant and evidently does not add clarity.

__22. $R_0 = 30$ seems high for measles (I do know that some studies have estimated it as that high or higher ...)__

Various time series investigations have found a preference for higher $R_0$ values for measles than some other traditional estimation methods, for reasons not yet fully understood.

__23. p. 35 italicize Leptospira interrogans?__

Done.

### ibpf vignette

__24. p.7 "provide[s] an adequate"__

* Done.

__25. p. 7 delete comma after "but"__

* Done.

__26. "Jensen's [inequality]"__

* Done.

__27. "This is [in] contrast"__

* Done.

__28. p. 9 "for one [parameter]"__

* Done.

__29. p. 12 "agree on the most problematic observation[s]" ?__

* Done.

__30. p. 14 "For our measles spatPomp [example]"__

* Done

__31. Fig 8: are the trace plots constant for the last parameters since these were held fixed in the example? Is it true that we shouldn't interpret the trace plots in quite the same way as for MCMC (since we are not trying to take a sample, i.e. we don't care about mixing or burn-in)?__

* Your explanations are correct. We have added a corresponding clarification to the IBPF tutorial.

__32. Wheeler et al; "Ionides, E. L." (not j.)"__

* Done.

### Other points

__33. plot method could/should add a y-axis label to the plot that indicates the `log10(x+1)` transformation? (Why not use `scale_y_log10()` [which automatically puts zero values (log(x) → -∞) at the edge of the plot] or the `log1p` transformation?__

* Thanks for these suggestions: I've adopted a `scale_y_log10()` solution using 'x+1', and labeling the axis accordingly.

__34. Titles of some help pages are noninformative (`runit_measure`, `munit_measure`)__

* Done.

__35. `plot` page headings could be improved__

* We have reworked the help page for `plot`.

__36. use `FALSE`, `TRUE` rather than `T`, `F`; clarify that `log=TRUE` uses `log(1+x)`__

* Done.
