### 2024-11-04 #7008 @ionides
Hi @gkthiruvathukal, I realized I did not address you in my response to Ben Bolker. Just to clarify, I have addressed his comments in the package, paper and tutorials. Please let me know if you need more from me at this time. Best, Ed.
