### 2024-11-07 #7008 @ionides
@gkthiruvathukal, this is taking a few days because we have decided it is time to move spatPomp to an organization (github.com/spatPomp-org/spatPomp rather than github.com/kidusasfaw/spatPomp). That should be done anyhow, since the original maintainer, @kidusasfaw, has less time for this project in his current job. The organization framework will be better for future growth, with a core development team that has published procedures (in https://github.com/spatPomp-org/.github/blob/main/profile/Governance.md) for incorporating new developers. This does not directly affect the software, though it does help respond to peer review comments about clarifying the process for developers to join the project.

I understand it adds a logistic issue with JOSS to move the location of the project repo at this time. If you have a strong preference for us to make the move after finishing the peer review process, we could do that. My assumption was that any improvements would be best done before finalizing the JOSS version. So, my working hypothesis is that we will get back to you when the spatPomp repo is fully installed at spatPomp-org and released from there to Zenodo. Please let me know if you recommend any different course of action.
