+++ DESCRIPTION
@@ -1,8 +1,8 @@
 Package: spatPomp
 Type: Package
 Title: Inference for Spatiotemporal Partially Observed Markov Processes
-Version: 0.37.0
-Date: 2024-10-25
+Version: 0.38.0
+Date: 2024-11-08
 Authors@R: c(
   person("Kidus", "Asfaw", email = "kidusasfaw1990@gmail.com", role = c("aut")),
   person("Edward", "Ionides", email = "ionides@umich.edu",role = c("cre","aut")),
@@ -13,13 +13,13 @@ Authors@R: c(
   person("Jifan", "Li", email = "jifanli@umich.edu", role = c("ctb")),
   person("Ning", "Ning", email = "patning@tamu.edu", role = c("ctb")),
   person("Haogao", "Gu", email = "hggu@connect.hku.hk", role = c("ctb")))
-URL: https://github.com/kidusasfaw/spatPomp
+URL: https://github.com/spatPomp-org/spatPomp
 Description: Inference on panel data using spatiotemporal partially-observed Markov process (SpatPOMP) models. The 'spatPomp' package extends 'pomp' to include algorithms taking advantage of the spatial structure in order to assist with handling high dimensional processes. See Asfaw et al. (2024) <doi:10.48550/arXiv.2101.01157> for further description of the package.
 SystemRequirements: For Windows users, Rtools (see https://cran.r-project.org/bin/windows/Rtools/).
 License: GPL-3
 Encoding: UTF-8
 LazyData: true
-BugReports: https://github.com/kidusasfaw/spatPomp/issues/
+BugReports: https://github.com/spatPomp-org/spatPomp/issues/
 Depends: pomp (>= 5.4), R(>= 4.1), methods
 LinkingTo: pomp
 Suggests: doParallel (>= 1.0.11), parallel, doRNG
