### 2024-11-08 1dd09b3
version 1.0.0 prepared for JOSS
