### 2024-11-08 1b7232b
update paper.md for relocation to spatPomp-org
