### 2024-11-08 #7008 @ionides
Hi @gkthiruvathukal, Here are the responses to the author tasks.

As I mentioned, the repository for the project has moved to
https://github.com/spatPomp-org/spatPomp
The current version of `paper.md` is on the joss branch of this repository. Please let me know if there's anything else I can do to help.

Best,
Ed

* Double check authors and affiliations (including ORCIDs)

Done.

* Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

Version 1.0.0

* Archive the release on Zenodo/figshare/etc and post the DOI here.

https://doi.org/10.5281/zenodo.14058236

* Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

Done.

* Make sure that the license listed for the archive is the same as the software license.

Done.
