+++ cran-comments.md
@@ -5,21 +5,23 @@ spatPomp 1.0.0 updates 0.36.1 with various minor changes requested during peer r
 
 #### Test environments
 
-Experimenting with test validation based on GitHub actions in
+Test validation based on GitHub actions in
 .github/workflows/r-cmd-check.yml
 and 
 .github/workflows/test-coverage.yml
 
 R-CMD-check / ubuntu-latest (devel) (push) Successful in 3m
-test-coverage / test-coverage (push) Successful in 5m
+test-coverage / test-coverage (push) Successful in 6m
 R-CMD-check / ubuntu-latest (release) (push) Successful in 3m
-R-CMD-check / ubuntu-20.04 (4.1.0) (push) Successful in 3m
+R-CMD-check / ubuntu-20.04 (4.1.0) (push) Successful in 4m
 R-CMD-check / macOS-latest (release) (push) Successful in 2m
-R-CMD-check / windows-latest (release) (push) Successful in 5m
-codecov/patch - Coverage not affected when comparing bc34025...bb41d65
-codecov/project - 100.00% (+0.08%) compared to bc34025
+R-CMD-check / windows-latest (release) (push) Successful in 6m
+R-CMD-check / ubuntu-latest (devel) (push) Successful in 3m
+
+covr unit test code coverage: 100%
 
 #### Reverse dependencies
 
 none found by devtools::revdep("spatPomp")
+
 
