+++ .github/workflows/binary-build.yml
@@ -3,7 +3,7 @@
 on:
   push:
     tags:
-      - '*.*.*'
+      - '*.*.*.*'
   workflow_dispatch:
 
 name: binary-build
