### 2024-11-16 #7008 @ionides
Hi @gkthiruvathukal,

A couple of updates:

1. spatPomp 1.0.0 is now on CRAN

2. I updated the joss and master branches on https://github.com/kidusasfaw/spatPomp to match https://github.com/spatPomp-org/spatPomp so it will not matter which site JOSS uses for the final version of the manuscript. Once this issue is resolved, we will archive https://github.com/kidusasfaw/spatPomp and continue development only on https://github.com/spatPomp-org/spatPomp.

Thanks,
Ed
