### 2024-11-25 #7008 @ionides
Hi  @gkthiruvathukal <https://github.com/gkthiruvathukal>, I'm not sure of
the procedure at this point, so I'm just checking in for confirmation that
I have completed everything on the author checklist (with responses in this
thread) so we are waiting for JOSS to get back to us with any further
requests. Thanks, Ed.

Edward L. Ionides
Associate Chair for Undergraduate Studies and Professor,
Department of Statistics, University of Michigan
1085 South University, Ann Arbor, MI 48109-1107
email: ***@***.***
phone: 734 615 3332
office: 453 West Hall

On Thu, Nov 7, 2024 at 1:12 PM George K. Thiruvathukal <
***@***.***> wrote:

> @ionides <https://github.com/ionides> I actually think this is a good
> idea to move it to an organization. I believe the system is designed to
> allow this change but don't know whether you or I would need to update the
> repository associated with the submission. I may have to read the manual.
> In any event, let me know when ready and I think we can easily figure this
> out.
>
