### 2024-11-26 #7008 @gkthiruvathukal
@ionides I set the version but you did not do a Release on GitHub. Your last version is 0.35. Can you please tag and release on GitHub, too? CRAN releases are not the same as a GitHub releases.
