### 2024-11-26 #7008 @ionides
@gkthiruvathukal, the 1.0.0 release is tagged on the new website for the organization at https://github.com/spatPomp-org/spatPomp. Perhaps you were being directed to the old website at https://github.com/kidusasfaw/spatPomp, which is indeed at 0.35. That repository is archived and no longer in development. It could be brought back to life and updated, if that is necessary. Alternatively, if this explanation is sufficient for you to keep moving, that would be great. Thanks, Ed.
