### 2024-12-02 #7008 @ionides
Hi @gkthiruvathukal, The old site at https://github.com/kidusasfaw/spatPomp is now given a matching 1.0.0 tag to the new site at https://github.com/spatPomp-org/spatPomp and the version on CRAN. Apologies for the delay with this - unarchiving the old site required coordination among the development team which is exactly the sort of thing that is smoothed over by moving to an organization. Best, Ed.
