### 2024-12-11 #7008 @ionides
Hi @gkthiruvathukal, I'm checking in on this issue. I think we fully addressed your previous request on Dec 2 so there is nothing else that my coauthors and I need to do at this point. But please let me know if there's anything else we can do to help. Best, Ed.
