### 2024-12-11 #7008 @gkthiruvathukal
Thanks, Ed @ionides. It looks like we are ready to move toward acceptance. I'm going to another read of the paper but need a bit more time as this is finals week at my university. I should be able to finish by the weekend.
