### 2024-12-20 #7008 @ionides
Hi @gkthiruvathukal, Thanks for your work on this. I attended to all the author checklist points in my response on Nov 8. On Nov 25, you confirmed these responses and implemented some of them (asking the editorial bot to set up the version and zenodo archive). I've re-checked the post-review checklist below, also mentioning the dates they were previously checked on https://github.com/openjournals/joss-reviews/issues/7008 in case that helps to track down any incompleteness or misunderstanding. Please let me know if you notice anything I have overlooked.

* Double check authors and affiliations (including ORCIDs)

Confirmed on 11/8 and reconfirmed here, on 12/20.

* Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

Done on 11/25. The tagged version 1.0.0 is archived on 10.5281/zenodo.14058236, and is on the current GitHub site, https://github.com/spatPomp-org/spatPomp, as well as the deprecated GitHub site, https://github.com/kidusasfaw/spatPomp, and the CRAN repository, https://cran.r-project.org/web/packages/spatPomp/index.html. This version was only updated on the deprecated site on 12/2, so that could have caused some difficulty.

* Archive the release on Zenodo/figshare/etc and post the DOI here.

Done, and added by JOSS on 11/25. The site is https://zenodo.org/records/14058236

* Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

Confirmed on 11/8, and reconfirmed here, on 12/20.

* Make sure that the license listed for the archive is the same as the software license.

Confirmed on 11/8, and reconfirmed here, on 12/20. Both are GNU GPL Version 3. I note that the JOSS manuscript is listed as CC BY 4.0. I assume that is the license for the manuscript rather than the software, and that these things do not have to be the same. My understanding is that Creative Commons is designed for published documents, whereas GPL is primarily designed for software, so this distinction seems to be appropriate. I don't see a problem here, but if that is what you have in mind then please let me know.
