+++ paper.bib
@@ -50,7 +50,7 @@
   journal = {Journal of Basic Engineering},
   volume = 82, 
   pages = {35--45},
-  doi={10.1115/1.366255}
+  doi={10.1115/1.3662552}
 }
 
 @Article{king16,
