### 2024-12-21 #7008 @ionides
Hi @gkthiruvathukal, I've fixed the broken doi on both the (identical) https://github.com/spatPomp-org/spatPomp and https://github.com/kidusasfaw/spatPomp versions of the JOSS paper. Glad to see that we're getting close! Thanks, Ed.
